연관관계 설정 및 paging 수정 #71

+++ src/main/java/com/travel/travel_project/domain/faq/FaqDTO.java
@@ -24,7 +24,7 @@ public class FaqDTO {
 
     @NotNull(message = "FAQ CODE 입력은 필수입니다.")
     @ApiModelProperty(required = true, value = "faq_code")
-    private Long faqCode;
+    private Integer faqCode;
 
     @NotEmpty(message = "FAQ 제목 입력은 필수입니다.")
     @ApiModelProperty(required = true, value = "title")
